added main, fixed double fields

+++ src/SaverchenkoGROUP10LAB1VARC2/Nutritious.java
@@ -1,5 +1,6 @@
 package SaverchenkoGROUP10LAB1VARC2;
 
-public interface Nutritious {
-    public abstract double calculateCalories();
+public interface Nutritious
+{
+    public abstract int calculateCalories();
 }
